# docs/swagger/index.yml
openapi: 3.0.3
info:
  title: online-shop-express
  version: 1.0.0
  contact: {}
servers:
  - url: 127.0.0.1
paths:
...
  /api/v1/verification_code:
    post:
      tags:
        - user-service
      summary: send-verification-code
      description: send-verification-code
      operationId: sendVerificationCode
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: rezaahmadi@gmail.com
                password:
                  type: string
                  example: '12345678'
            examples:
              send-verification-code:
                value:
                  email: rezaahmadi@gmail.com
                  password: '12345678'
      responses:
        '200':
          description: ''
  /api/v1/verify:
    post:
      tags:
        - user-service
      summary: verify-account
      description: verify-account
      operationId: verifyAccount
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: rezaahmadi2@gmail.com
                password:
                  type: string
                  example: '12345678'
                verificationCode:
                  type: string
                  example: '970175'
            examples:
              verify-account:
                value:
                  email: rezaahmadi2@gmail.com
                  password: '12345678'
                  verificationCode: '970175'
      responses:
        '200':
          description: ''
  /api/v1/disable_account:
    post:
      tags:
        - user-service
      summary: disable-account
      description: disable-account
      operationId: disableAccount
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: alibasavqwad138193@gmail.com
                password:
                  type: string
                  example: '12345678'
            examples:
              disable-account:
                value:
                  email: alibasavqwad138193@gmail.com
                  password: '12345678'
      responses:
        '200':
          description: ''
...
  /api/v1/profile:
    get:
      tags:
        - user-service
      summary: read-profile
      description: read-profile
      operationId: readProfile
      requestBody:
        content:
          application/json:
            examples:
              read-profile:
                value: ''
      responses:
        '200':
          description: ''
      security:
        - bearerAuth: []
    patch:
      tags:
        - user-service
      summary: update-profile
      description: update-profile
      operationId: updateProfile
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                  example: Mahdi
                lastName:
                  type: string
                  example: Jokar
                phoneNumber:
                  type: string
                  example: '09382790111'
            examples:
              update-profile:
                value:
                  firstName: Mahdi
                  lastName: Jokar
                  phoneNumber: '09382790111'
      responses:
        '200':
          description: ''
      security:
        - bearerAuth: []
...
  /api/v1/generate_token/:
    post:
      tags:
        - user-service
      summary: generate-access-token
      description: generate-access-token
      operationId: generateAccessToken
      responses:
        '200':
          description: ''
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...